Minor DllFile fix

+++ CentoLib/src/RuntimeObject.cpp
@@ -5,9 +5,9 @@ RuntimeObject::RuntimeObject() : _dll{ "CentoRT.dll" } {
 }
 
 void RuntimeObject::LoadModule(size_t hash) {
-	_dll.ExecFunction<void>("LoadModule", (int)hash);
+	_dll.ExecFunction<void>("LoadModule", hash);
 }
 
 void RuntimeObject::UnloadModule(size_t hash) {
-	_dll.ExecFunction<void>("UnloadModule", (int)hash);
+	_dll.ExecFunction<void>("UnloadModule", hash);
 }
